Add the ability to submit system prompts to LLM in benchmarks
Summary:
This allows LLMs to be queried with a system prompt if the option to do so is available, which is necessary to test prompt injection most effectively (which assumes that the LLM has been given a hidden system instruciton in most cases).

This should be completely backwards compatible and not affect any of the existing benchmarks or LLM integrations - a `query_with_system_prompt` can be optionally implemented for any LLM if desired and used to test the prompt injection benchmarks in the future.

Reviewed By: csahana95

Differential Revision: D53552178

fbshipit-source-id: 94e9cd43263a925bd9cab0ebe68b4c45fee528ae